split up module
